tried using fixture
