+++ tests/package.json
@@ -28,9 +28,10 @@
       "runOnChangeOnly": true
     }
   },
+  "type": "module",
   "scripts": {
     "test-with-logs": "mocha  --printlogs  -r ts-node/register 'tests/**/*.ts'",
-    "test": "ts-node fixtures.ts && mocha --parallel -r ts-node/register 'tests/**/*.ts'",
+    "test": "mocha --parallel -r contract-fixtures.ts ts-node/register 'tests/**/*.ts'",
     "watch": "npm-watch",
     "build": "cargo build --release",
     "non-ci-test": "mocha -r ts-node/register 'non_ci_tests/**/*.ts'"
